Fix peerDependencies edge case
introduced by firebase. I'm not even sure if this is *supposed* to work
but it does so *shrug*. So firebase depends on @firebase/messaging which
has peerDependencies for @firebase/app and @firebase/app-types. However
firebase itself only depends on @firebase/app which in turn depends on
@firebase/app-types. This previously didn't work because the the
conditional expected both peerDependencies to be served by the requirer
directly. If pnpm (which is supposed to be strict) allows this then who
am I to argue? 😬